changing ACES to ACES conversion methods to use float

+++ com.unity.render-pipelines.core/ShaderLibrary/ACES.hlsl
@@ -265,7 +265,7 @@ half3 ACEScc_to_ACES(half3 x)
 // converts ACES2065-1 (AP0 w/ linear encoding) to
 //          ACEScg (AP1 w/ linear encoding)
 //
-half3 ACES_to_ACEScg(half3 x)
+float3 ACES_to_ACEScg(float3 x)
 {
     return mul(AP0_2_AP1_MAT, x);
 }
@@ -276,7 +276,7 @@ half3 ACES_to_ACEScg(half3 x)
 // converts ACEScg (AP1 w/ linear encoding) to
 //          ACES2065-1 (AP0 w/ linear encoding)
 //
-half3 ACEScg_to_ACES(half3 x)
+float3 ACEScg_to_ACES(float3 x)
 {
     return mul(AP1_2_AP0_MAT, x);
 }
